@@ -1,16 +1,18 @@
 <template>
-	<div class="Home">
-		<Banner />
-	</div>
+  <div class="home">
+    <img alt="Vue logo" src="../assets/logo.png">
+    <HelloWorld msg="Welcome to Your Vue.js App"/>
+  </div>
 </template>
 
 <script>
-import Banner from '@/components/Banner.vue';
+// @ is an alias to /src
+import HelloWorld from '@/components/HelloWorld.vue'
 
 export default {
-	name: 'Home',
-	components: {
-		Banner,
-	},
-};
+  name: 'Home',
+  components: {
+    HelloWorld
+  }
+}
 </script>
